backend/requirements.txt
Flask==2.3.2
pymongo==4.3.3
openai==0.28.0
flask-cors==4.0.0
python-dotenv==1.0.0
numpy==1.26.4
scikit-learn==1.4.2